Move location selection screen to landing page
